+++ lib/test.dart
@@ -24,6 +24,7 @@ export 'src/frontend/future_matchers.dart';
 export 'src/frontend/on_platform.dart';
 export 'src/frontend/prints_matcher.dart';
 export 'src/frontend/skip.dart';
+export 'src/frontend/spawn_hybrid.dart';
 export 'src/frontend/tags.dart';
 export 'src/frontend/test_on.dart';
 export 'src/frontend/throws_matcher.dart';